# openapi/generated_openapi.json
{
  "components": {
    "schemas": {
      "AIProfileActionResponse": {
        "properties": {
          "ai_profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Profile"
          },
          "ai_profile_draft": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Profile Draft"
          },
          "ai_profile_status": {
            "default": "none",
            "title": "Ai Profile Status",
            "type": "string"
          },
          "ai_profile_updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Profile Updated At"
          },
          "profile_id": {
            "title": "Profile Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "generated",
              "approved",
              "rejected"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "profile_id"
        ],
        "title": "AIProfileActionResponse",
        "type": "object"
      },
...
      "DraftUpdateRequest": {
        "description": "Update draft fields \u2014 validated dynamically against config schema.",
        "properties": {
          "fields": {
            "additionalProperties": true,
            "title": "Fields",
            "type": "object"
          }
        },
        "required": [
          "fields"
        ],
        "title": "DraftUpdateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Connecting specialty grain producers with global buyers",
    "title": "GrainPlaza",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/roles/buyer/draft": {
      "get": {
        "operationId": "get_draft_buyer_api_roles_buyer_draft_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "title": "Session Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Draft Buyer",
        "tags": [
          "generated-roles"
        ]
      },
      "put": {
        "operationId": "update_draft_buyer_api_roles_buyer_draft_put",
        "parameters": [
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "title": "Session Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Draft Buyer",
        "tags": [
          "generated-roles"
        ]
      }
    },
    "/api/roles/buyer/draft/submit": {
      "post": {
        "operationId": "submit_draft_buyer_api_roles_buyer_draft_submit_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "title": "Session Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Draft Buyer",
        "tags": [
          "generated-roles"
        ]
      }
    },
    "/api/roles/buyer/me": {
      "get": {
        "operationId": "me_buyer_api_roles_buyer_me_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "title": "Session Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Me Buyer",
        "tags": [
          "generated-roles"
        ]
      }
    },
    "/api/roles/buyer/register": {
      "post": {
        "operationId": "register_buyer_api_roles_buyer_register_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "title": "Session Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Buyer",
        "tags": [
          "generated-roles"
        ]
      }
    },
    "/api/roles/buyer/{profile_id}": {
      "get": {
        "operationId": "get_profile_buyer_api_roles_buyer__profile_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "title": "Session Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Profile Buyer",
        "tags": [
          "generated-roles"
        ]
      },
      "put": {
        "operationId": "update_profile_buyer_api_roles_buyer__profile_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "title": "Session Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Profile Buyer",
        "tags": [
          "generated-roles"
        ]
      }
    },
    "/api/roles/buyer/{profile_id}/ai-approve": {
      "post": {
        "operationId": "ai_approve_buyer_api_roles_buyer__profile_id__ai_approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "title": "Session Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIProfileActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ai Approve Buyer",
        "tags": [
          "generated-roles"
        ]
      }
    },
    "/api/roles/buyer/{profile_id}/ai-generate": {
      "post": {
        "operationId": "ai_generate_buyer_api_roles_buyer__profile_id__ai_generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "title": "Session Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIProfileActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ai Generate Buyer",
        "tags": [
          "generated-roles"
        ]
      }
    },
    "/api/roles/buyer/{profile_id}/ai-reject": {
      "post": {
        "operationId": "ai_reject_buyer_api_roles_buyer__profile_id__ai_reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "title": "Session Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIProfileActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ai Reject Buyer",
        "tags": [
          "generated-roles"
        ]
      }
    },
    "/api/roles/producer/draft": {
      "get": {
        "operationId": "get_draft_producer_api_roles_producer_draft_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "title": "Session Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Draft Producer",
        "tags": [
          "generated-roles"
        ]
      },
      "put": {
        "operationId": "update_draft_producer_api_roles_producer_draft_put",
        "parameters": [
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "title": "Session Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Draft Producer",
        "tags": [
          "generated-roles"
        ]
      }
    },
    "/api/roles/producer/draft/submit": {
      "post": {
        "operationId": "submit_draft_producer_api_roles_producer_draft_submit_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "title": "Session Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Draft Producer",
        "tags": [
          "generated-roles"
        ]
      }
    },
    "/api/roles/producer/me": {
      "get": {
        "operationId": "me_producer_api_roles_producer_me_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "title": "Session Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Me Producer",
        "tags": [
          "generated-roles"
        ]
      }
    },
    "/api/roles/producer/register": {
      "post": {
        "operationId": "register_producer_api_roles_producer_register_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "title": "Session Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Producer",
        "tags": [
          "generated-roles"
        ]
      }
    },
    "/api/roles/producer/{profile_id}": {
      "get": {
        "operationId": "get_profile_producer_api_roles_producer__profile_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "title": "Session Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Profile Producer",
        "tags": [
          "generated-roles"
        ]
      },
      "put": {
        "operationId": "update_profile_producer_api_roles_producer__profile_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "title": "Session Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Profile Producer",
        "tags": [
          "generated-roles"
        ]
      }
    },
    "/api/roles/producer/{profile_id}/ai-approve": {
      "post": {
        "operationId": "ai_approve_producer_api_roles_producer__profile_id__ai_approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "title": "Session Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIProfileActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ai Approve Producer",
        "tags": [
          "generated-roles"
        ]
      }
    },
    "/api/roles/producer/{profile_id}/ai-generate": {
      "post": {
        "operationId": "ai_generate_producer_api_roles_producer__profile_id__ai_generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "title": "Session Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIProfileActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ai Generate Producer",
        "tags": [
          "generated-roles"
        ]
      }
    },
    "/api/roles/producer/{profile_id}/ai-reject": {
      "post": {
        "operationId": "ai_reject_producer_api_roles_producer__profile_id__ai_reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "title": "Session Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIProfileActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ai Reject Producer",
        "tags": [
          "generated-roles"
        ]
      }
    },
...
  }
}